linebot_service/requirements.txt
Flask==2.1.1
line-bot-sdk==1.20.0
Werkzeug==2.0.3
pytz
requests

 